rename test_custom_hook
