+++ src/components/datagrid/utils/scrolling.test.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { testCustomHook } from '../../../test';
+import { testCustomHook } from '../../../test/test_custom_hook.test_helper';
 import { useScrollCellIntoView } from './scrolling';
 
 // see scrolling.spec.tsx for E2E useScroll tests
