+++ src/test/index.ts
@@ -9,7 +9,6 @@
 export { requiredProps } from './required_props';
 export { takeMountedSnapshot } from './take_mounted_snapshot';
 export { findTestSubject } from './find_test_subject';
-export { testCustomHook } from './test_custom_hook';
 export {
   startThrowingReactWarnings,
   stopThrowingReactWarnings,
